drivers/usb/serial: move cp2102 to a subdirectory
