+++ drivers/usb/devices/serial/meson.build
@@ -2,7 +2,7 @@ posix_bragi = cxxbragi.process(protos/'posix/posix.bragi')
 
 src = [
 	'src/main.cpp',
-	'src/cp2102.cpp',
+	'src/cp2102/cp2102.cpp',
 	posix_bragi,
 ]
